requirements.txt
python==3.8.13
django==4.1
django-debug-toolbar==3.7.0
pymysql==1.0.2
djangorestframework==3.14.0
django-imagekit==4.1.0
pillow==9.3.0
sympy==1.11.1
django-rest-knox==4.2.0
django-filter==22.1
requests==2.28.1
django-rest-passwordreset==1.3.0
